+++ Changelog.md
@@ -2,3 +2,4 @@
 
 ### Version 1.16.5 - 1.5.0
  - add forgotten well structure
+ - add jungle temple structure