+++ Changelog.md
@@ -1,1 +1,8 @@
 # Changelog
+
+### Version 1.16.5 - 1.5.0
+ - add forgotten well structure
+ - add jungle temple structure
+ - add underground temple structure
+ - moved config to normal config folder
+